https://www.hackerrank.com/challenges/crush
Algorithmic Crush
https://www.hackerrank.com/challenges/crush

You are given a list of size N, initialized with zeroes. You have to perform M operations on the list and output the maximum of final values of all the  elements in the list. For every operation, you are given three integers a, b and k and you have to add value k to all the elements ranging from index a to b (both inclusive).

Constraints:
3 ≤ N ≤ 10^7
1 ≤ M ≤ 2*10^5
1 ≤ a ≤ b ≤ N
0 ≤ k ≤ 10^9