https://www.hackerrank.com/challenges/insert-a-node-at-the-tail-of-a-linked-list
Insert a Node at the Tail of a Linked List 

https://www.hackerrank.com/challenges/insert-a-node-at-the-tail-of-a-linked-list

You are given the pointer to the head node of a linked list and an integer to add to the list. Create a new node with the given integer. Insert this node at the tail of the linked list and return the head node. The given head pointer may be null, meaning that the initial list is empty.