https://www.hackerrank.com/challenges/insert-a-node-at-the-head-of-a-linked-list
Insert a node at the head of a linked list 

https://www.hackerrank.com/challenges/insert-a-node-at-the-head-of-a-linked-list

You’re given the pointer to the head node of a linked list and an integer to add to the list. Create a new node with the given integer, insert this node at the head of the linked list and return the new head node. The head pointer given may be null meaning that the initial list is empty.