https://www.hackerrank.com/challenges/delete-a-node-from-a-linked-list
Delete a Node 

https://www.hackerrank.com/challenges/delete-a-node-from-a-linked-list

You’re given the pointer to the head node of a linked list and the position of a node to delete. Delete the node at the given position and return the head node. A position of 0 indicates head, a position of 1 indicates one node away from the head and so on. The list may become empty after you delete the node.