https://www.hackerrank.com/challenges/get-the-value-of-the-node-at-a-specific-position-from-the-tail
Get Node Value 

https://www.hackerrank.com/challenges/get-the-value-of-the-node-at-a-specific-position-from-the-tail

You’re given the pointer to the head node of a linked list and a specific position. Counting backwards from the tail node of the linked list, get the value of the node at the given position. A position of 0 corresponds to the tail, 1 corresponds to the node before the tail and so on.