https://www.hackerrank.com/challenges/delete-duplicate-value-nodes-from-a-sorted-linked-list
Delete duplicate-value nodes from a sorted linked list

https://www.hackerrank.com/challenges/delete-duplicate-value-nodes-from-a-sorted-linked-list

You're given the pointer to the head node of a sorted linked list, where the data in the nodes is in ascending order. Delete as few nodes as possible so that the list does not contain any value more than once. The given head pointer may be null indicating that the list is empty.