https://www.hackerrank.com/challenges/binary-search-tree-insertion
Binary Search Tree : Insertion
https://www.hackerrank.com/challenges/binary-search-tree-insertion

You are given a pointer to the root of a binary search tree and a value to be inserted into the tree. Insert this value into its appropriate position in the binary search tree and return the root of the updated binary tree. You just have to complete the function.